New editor (#31)
* New sensitivity analysis

* Saving work in progress

* Adding the type attribute to the Text dataclass. Other fixes

* Being able to filter steps using parameters.
  Being able to parse (-inf,0dB) in complex values
  Fixing problem on the run_analysis() on the quick_sensitivity_analysis.py where the order of the components 
  was not being considered.

* Moving Schematic related classes and methods out of the BaseEditor into a BaseSchematic subclass
  Fixes on the get_text_space() method on AscEditor

* Fixes on the add_instruction() method of AscEditor class. It now correctly replaces unique instructions.

* Fixes in the test_asc_editor.py testbench

* Implementing an add_sources to the client-server implementation.

* Timeout is no longer fixed on SimServer. In can be overrided on the class constructor. This was also added to 
  the run_server.py.

* Avoiding closing the session 2 times.

* Improving the debug messages on the sim_server.py

* Corrections on the toolkit:
   - SimAnalysis.run() doesn't need the netlist as parameter.
   - ToleranceDeviations.run() should not be used. Now raises a message advising the usage of either the 
     run_testbench() or run_analysis()

* Adding missing methods into the AscEditor: get_component_position() and set_component_position()
  Fixes on the SpiceCircuit:
  - Timeout defaults to None
  - get_component_line() method didn't exist. Replaced with _get_line_starting_with()

* Removing the run() implementation on the ToleranceDeviations class

* moving Commmand Line Interface (CLI) scripts to its own directory : cli

* Adding create_blank parameter to the reset_netlist() and aligning all subclasses to this.

* Adding the add_component() method.

* Refactoring all the Editors to use the same class structure. This is done with the purpose of being able to 
  export files from one schematic type to another.

*  Adding a tool to convert from ASC files to QSCH Files.

* Adding missing file for testbench

* Suporting the testbench montecarlo simulation in Qspice.

* Supporting other separators on the histogram script.
Correcting the problem where the script was being executed twice.
Using try_convert_value() so that other formats can be used.

* Adding a script that includes python code in the readme.md

* Corrections made after executing all examples.

* Adding the missing reset_netlist() onto the SpiceCircuit Class

+++ spicelib/sim/sim_runner.py
@@ -90,7 +90,7 @@ The RAW file and the LOG file names. Below is an example of a callback function:
     def processing_data(raw_filename, log_filename):
         '''This is a call back function that just prints the filenames'''
         print("Simulation Raw file is %s. The log is %s" % (raw_filename, log_filename)
-        # Other code below either using LTSteps.py or raw_read.py
+        # Other code below either using ltsteps.py or raw_read.py
         log_info = LTSpiceLogReader(log_filename)
         log_info.read_measures()
         rise, measures = log_info.dataset["rise_time"]
@@ -538,7 +538,7 @@ class SimRunner(object):
         :type workfile: Path
         :return: Nothing
         """
-        if workfile.exists():
+        if workfile is not None and workfile.exists():
             _logger.info("Deleting..." + workfile.name)
             workfile.unlink()
